+++ R/6-fit_mod.R
@@ -547,6 +547,7 @@ fit_mod <-
     data_list_reorganized$linkage_link         <- .linkage_enc$linkage_link
     data_list_reorganized$linkage_re_index     <- .linkage_enc$linkage_re_index
     data_list_reorganized$linkage_re_sigma      <- .linkage_enc$linkage_re_sigma
+    data_list_reorganized$linkage_re_struct     <- .linkage_enc$linkage_re_struct
     data_list_reorganized$linkage_re_sigma_prior_family <- .linkage_enc$linkage_re_sigma_prior_family
     data_list_reorganized$linkage_re_sigma_prior_p1     <- .linkage_enc$linkage_re_sigma_prior_p1
     data_list_reorganized$linkage_re_sigma_prior_p2     <- .linkage_enc$linkage_re_sigma_prior_p2